split pkg/osquery/runtime/runtime.go into separate files for osquery instance and runner
